--- dev-kit/app/Config/Database.ts
@@ -1,22 +1,0 @@
-import path from "path";
-import { IDatabaseConfig } from "../../../index";
-
-const config: IDatabaseConfig = {
-  client: process.env.DB_CLIENT || "mysql",
-  connection: {
-    host: process.env.DB_HOST || "localhost",
-    user: process.env.DB_USER || "user",
-    password: process.env.DB_PASSWORD || "password",
-    database: process.env.DB_DATABASE || "database",
-    filename: path.join(__dirname, "..", "..", "mydb.sqlite"),
-  },
-  pool: {
-    min: 2,
-    max: 10,
-  },
-  migrations: {
-    tableName: "knex_migrations",
-  },
-};
-
-export default config;
