+++ dev-kit/app/config.ts
@@ -1,0 +1,28 @@
+import path from "path";
+import { LogLevels, IApplicationConfig } from "../../index";
+
+const config: IApplicationConfig = {
+  prefix: "api",
+  env: process.env.NODE_ENV || "production",
+  port: process.env.APP_PORT ? parseInt(process.env.APP_PORT) : 3000,
+  logLevel: LogLevels.INFO,
+  database: {
+    client: process.env.DB_CLIENT || "mysql",
+    connection: {
+      host: process.env.DB_HOST || "localhost",
+      user: process.env.DB_USER || "user",
+      password: process.env.DB_PASSWORD || "password",
+      database: process.env.DB_DATABASE || "database",
+      filename: path.join(__dirname, "..", "..", "mydb.sqlite"),
+    },
+    pool: {
+      min: 2,
+      max: 10,
+    },
+    migrations: {
+      tableName: "knex_migrations",
+    },
+  },
+};
+
+export default config;
