+++ dev-kit/app/v1/Events/UserEvent.ts
@@ -1,4 +1,4 @@
-import { IHookParameter } from "../../../index";
+import { IHookParameter } from "../../../../index";
 
 const onAfterInsert = async ({ item }: IHookParameter) => {
   // console.log("onAfterInsert", item);
