+++ dev-kit/app/v1/config.ts
@@ -1,0 +1,10 @@
+import { IVersionConfig } from "../../../index";
+
+const config: IVersionConfig = {
+  transaction: [],
+  serializers: [],
+  supportedLanguages: ["en"],
+  defaultLanguage: "en",
+};
+
+export default config;
